vestigekv: native-dtype tensor-core dots in the scan, 6.6x kernel speedup
The convert-to-fp32 + ieee form ran the dot on CUDA cores: after storage
dropped to 16-bit the scan became compute-bound and LOST the bandwidth
win (serving scan 0.53 -> 1.28 ms/step at 256k; standalone 538 GB/s
effective vs the 1186 GB/s fp32 baseline). Native bf16/fp16 dots with
fp32 accumulation run on tensor cores at 3601 GB/s effective (0.018 ms
standalone, 6.6x) and are precision-sound: every bf16/fp16 product is
exact in the fp32 accumulator (mantissas square under 24 bits), so unlike
tf32 there is no silent truncation -- only accumulation ORDER differs.
Query operands are rounded to the storage dtypes at every scoring site
(kernels, eager paths, batched pack, and build's zp calibration), keeping
quantize-then-calibrate coherent.

Test contract updated: the eager reference now upcasts to fp32 BEFORE its
matmuls -- a native fp16 torch matmul uses reduced-precision reduction
(~1e-3) and the reference, not the kernel, flipped rows that fp64 says
must not fire. Kernel-vs-eager allows only boundary-row flips (score
within float rounding of the threshold); batched-vs-per-pair stays bit-
identical. Fire-set gate vs the all-fp32 reference: Jaccard mean 0.9957,
coverage 0.9865 vs 0.9863, zp recalibrated 4.5382 -> 4.5412. 71/71 tests.

+++ python/sglang/srt/layers/attention/vestigekv/recall_tier.py
@@ -213,9 +213,12 @@ class RecallTier:
             tgt_side = self.side[pa]  # [n_hard, 64]
             tgt_csk = self.csk[pa]  # [n_hard, r]
             tgt_rho = self.rho[pa]  # [n_hard]
-            idxs_t = (qh[:, D.KV_LORA_RANK :] * tgt_side).sum(-1) + (
-                qskh * tgt_csk
-            ).sum(-1)
+            # Same rounding as the serve-time scoring path: query operands
+            # rounded to the storage dtypes before the (exact-in-fp32)
+            # products, so zp is calibrated on exactly what the kernel scores.
+            idxs_t = (
+                qh[:, D.KV_LORA_RANK :].to(torch.bfloat16).float() * tgt_side.float()
+            ).sum(-1) + (qskh.half().float() * tgt_csk.float()).sum(-1)
             idxs_t = idxs_t * sc_
             cert_t = (
                 qresh * tgt_rho * sc_ / (D.KV_LORA_RANK - self.r) ** 0.5
@@ -276,8 +279,11 @@ class RecallTier:
         qres = (qe[:, : D.KV_LORA_RANK] - qsk @ self.V).norm(dim=-1)
         if self._qside_t is None:
             H = qe.shape[0]
-            self._qside_t = qe.new_empty(D.SIDECAR_DIM, H)
-            self._qsk_t = qe.new_empty(self.r, H)
+            # Storage dtypes, matching the kernel's native-dtype dots: bf16 for
+            # the sidecar branch (exact -- qbuf is bf16), fp16 for the sketch
+            # projection (rounded here AND at calibration, so zp covers it).
+            self._qside_t = qe.new_empty(D.SIDECAR_DIM, H, dtype=torch.bfloat16)
+            self._qsk_t = qe.new_empty(self.r, H, dtype=torch.float16)
             self._hit_buf = torch.empty(
                 self.side.shape[0], dtype=torch.int32, device=qe.device
             )
@@ -290,9 +296,9 @@ class RecallTier:
             # the score matrix the top-j selection ranks. The cap is explicit
             # opt-in; the default (uncapped) path is the fused one.
             idxs = (
-            qe[:, D.KV_LORA_RANK :] @ self.side.float().T
-            + qsk @ self.csk.float().T
-        ) * sc_
+                (qe[:, D.KV_LORA_RANK :].to(torch.bfloat16) @ self.side.T).float()
+                + (qsk.half() @ self.csk.T).float()
+            ) * sc_
             cert = (
                 (qres[:, None] * self.rho[None, :])
                 * sc_
@@ -368,8 +374,8 @@ class RecallTier:
         qsk = qe[:, : D.KV_LORA_RANK] @ self.V.T
         qres = (qe[:, : D.KV_LORA_RANK] - qsk @ self.V).norm(dim=-1)
         idxs = (
-            qe[:, D.KV_LORA_RANK :] @ self.side.float().T
-            + qsk @ self.csk.float().T
+            (qe[:, D.KV_LORA_RANK :].to(torch.bfloat16) @ self.side.T).float()
+            + (qsk.half() @ self.csk.T).float()
         ) * sc_
         cert = (
             (qres[:, None] * self.rho[None, :]) * sc_ / (D.KV_LORA_RANK - self.r) ** 0.5
